move logging into separate folder and add picologging handler
